feat: LRD load application(RDCC-1867)
